airflow/requirements.txt
apache-airflow-providers-google
google-api-python-client==2.87.0
pandas
pandas_gbq